+++ src/ValidationException.js
@@ -1,0 +1,7 @@
+export default class ValidationException extends Error {
+    static TAG = 'VIDMOM ERROR';
+    constructor(message) {
+        super(`${ValidationException.TAG}: ${message}`);
+        this.name = 'ValidationError';
+    }
+}
